// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const EnumOrderStatus = {
  PENDING: 'PENDING',
  PAYED: 'PAYED'
} as const

export type EnumOrderStatus = (typeof EnumOrderStatus)[keyof typeof EnumOrderStatus]


export const EnumUserRole = {
  ADMIN: 'ADMIN',
  SALER: 'SALER',
  USER: 'USER'
} as const

export type EnumUserRole = (typeof EnumUserRole)[keyof typeof EnumUserRole]